@@ -4,8 +4,8 @@ description: Flutter package to render html as widgets that supports hyperlink, 
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=3.3.0"
-  sdk: ">=2.18.0 <3.0.0"
+  flutter: ">=3.7.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
   flutter:
