@@ -1,9 +1,3 @@
-import 'dart:typed_data';
-
 import 'package:flutter_svg/flutter_svg.dart';
 
-PictureProvider? filePicture(
-  PictureInfoDecoderBuilder<Uint8List> decoderBuilder,
-  String path,
-) =>
-    null;
+BytesLoader? fileLoader(String path) => null;
