@@ -20,85 +20,67 @@ mixin SvgFactory on WidgetFactory {
   Widget? buildImageWidget(BuildMetadata meta, ImageSource src) {
     final url = src.url;
 
-    PictureProvider? provider;
+    BytesLoader? bytesLoader;
     if (url.startsWith('data:image/svg+xml')) {
-      provider = imageSvgFromDataUri(url);
+      bytesLoader = imageSvgFromDataUri(url);
     } else if (Uri.tryParse(url)?.path.toLowerCase().endsWith('.svg') == true) {
       if (url.startsWith('asset:')) {
-        provider = imageSvgFromAsset(url);
+        bytesLoader = imageSvgFromAsset(url);
       } else if (url.startsWith('file:')) {
-        provider = imageSvgFromFileUri(url);
+        bytesLoader = imageSvgFromFileUri(url);
       } else {
-        provider = imageSvgFromNetwork(url);
+        bytesLoader = imageSvgFromNetwork(url);
       }
     }
 
-    if (provider == null) {
+    if (bytesLoader == null) {
       return super.buildImageWidget(meta, src);
     }
 
-    return _buildSvgPicture(meta, src, provider);
+    return _buildSvgPicture(meta, src, bytesLoader);
   }
 
-  /// Returns an [ExactAssetPicture].
-  PictureProvider? imageSvgFromAsset(String url) {
+  /// Returns an [SvgAssetLoader].
+  BytesLoader? imageSvgFromAsset(String url) {
     final uri = Uri.parse(url);
     final assetName = uri.path;
     if (assetName.isEmpty) {
       return null;
     }
 
-    return ExactAssetPicture(
-      svgAllowDrawingOutsideViewBox
-          ? SvgPicture.svgStringDecoderOutsideViewBoxBuilder
-          : SvgPicture.svgStringDecoderBuilder,
+    return SvgAssetLoader(
       assetName,
-      package: uri.queryParameters['package'],
+      packageName: uri.queryParameters['package'],
     );
   }
 
-  /// Returns a [MemoryPicture].
-  PictureProvider? imageSvgFromDataUri(String dataUri) {
+  /// Returns a [SvgBytesLoader].
+  BytesLoader? imageSvgFromDataUri(String dataUri) {
     final bytes = bytesFromDataUri(dataUri);
     if (bytes == null) {
       return null;
     }
 
-    return MemoryPicture(
-      svgAllowDrawingOutsideViewBox
-          ? SvgPicture.svgByteDecoderOutsideViewBoxBuilder
-          : SvgPicture.svgByteDecoderBuilder,
-      bytes,
-    );
+    return SvgBytesLoader(bytes);
   }
 
-  /// Returns a [FilePicture].
-  PictureProvider? imageSvgFromFileUri(String url) {
+  /// Returns a [SvgFileLoader].
+  BytesLoader? imageSvgFromFileUri(String url) {
     final filePath = Uri.parse(url).toFilePath();
     if (filePath.isEmpty) {
       return null;
     }
 
-    return filePicture(
-      svgAllowDrawingOutsideViewBox
-          ? SvgPicture.svgByteDecoderOutsideViewBoxBuilder
-          : SvgPicture.svgByteDecoderBuilder,
-      filePath,
-    );
+    return fileLoader(filePath);
   }
 
-  /// Returns a [NetworkPicture].
-  PictureProvider? imageSvgFromNetwork(String url) {
+  /// Returns a [SvgNetworkLoader].
+  BytesLoader? imageSvgFromNetwork(String url) {
     if (url.isEmpty) {
       return null;
     }
 
-    return NetworkPicture(
-      svgAllowDrawingOutsideViewBox
-          ? SvgPicture.svgByteDecoderOutsideViewBoxBuilder
-          : SvgPicture.svgByteDecoderBuilder,
-      url,
-    );
+    return SvgNetworkLoader(url);
   }
 
   @override
@@ -107,13 +89,8 @@ mixin SvgFactory on WidgetFactory {
       case 'svg':
         _tagSvg ??= BuildOp(
           onWidgets: (meta, widgets) {
-            final provider = StringPicture(
-              svgAllowDrawingOutsideViewBox
-                  ? SvgPicture.svgStringDecoderOutsideViewBoxBuilder
-                  : SvgPicture.svgStringDecoderBuilder,
-              meta.element.outerHtml,
-            );
-            return [_buildSvgPicture(meta, const ImageSource(''), provider)];
+            final bytesLoader = SvgStringLoader(meta.element.outerHtml);
+            return [_buildSvgPicture(meta, const ImageSource(''), bytesLoader)];
           },
         );
         meta.register(_tagSvg!);
@@ -126,13 +103,13 @@ mixin SvgFactory on WidgetFactory {
   Widget _buildSvgPicture(
     BuildMetadata meta,
     ImageSource src,
-    PictureProvider provider,
+    BytesLoader bytesLoader,
   ) {
     final image = src.image;
     final semanticLabel = image?.alt ?? image?.title;
 
     return SvgPicture(
-      provider,
+      bytesLoader,
       allowDrawingOutsideViewBox: svgAllowDrawingOutsideViewBox,
       excludeFromSemantics: semanticLabel == null,
       fit: BoxFit.fill,
